Fix OpenOCD builder

+++ deb-openocd.sh
@@ -4,8 +4,11 @@ export VERSION=0.7.0
 export DEBVERSION=${VERSION}-1
 export URL="http://downloads.sourceforge.net/project/openocd/openocd/0.7.0/openocd-0.7.0.tar.bz2?r=http%3A%2F%2Fsourceforge.net%2Fprojects%2Fopenocd%2Ffiles%2Fopenocd%2F0.7.0%2F&ts=1368821905&use_mirror=switch"
 #Download it
-wget "$URL" -O ${NAME}_${VERSION}.orig.tar.bz2
-tar xzvf ${NAME}_${VERSION}.orig.tar.bz2
+if [ ! -f "${NAME}_${VERSION}.orig.tar.bz2" ]
+then
+    wget "$URL" -O ${NAME}_${VERSION}.orig.tar.bz2
+fi
+tar xjvf ${NAME}_${VERSION}.orig.tar.bz2
 cd openocd-${VERSION}
 rm -rf debian
 mkdir -p debian
@@ -36,11 +39,11 @@ echo -e '\tdh $@' >> debian/rules
 echo 'override_dh_auto_configure:' >> debian/rules
 echo -e "\t./configure --prefix=`pwd`/debian/${NAME}/usr" >> debian/rules
 echo 'override_dh_auto_build:' >> debian/rules
-echo -e '\tmake' >> debian/rules
+echo -e '\tmake -j8' >> debian/rules
 echo 'override_dh_auto_install:' >> debian/rules
 echo -e '\tmake install' >> debian/rules
 echo -e "\tmkdir -p debian/$NAME-dev/usr" >> debian/rules
-echo -e "\tmv debian/$NAME/usr/include debian/$NAME-dev/usr/" >> debian/rules
+#echo -e "\tmv debian/$NAME/usr/include debian/$NAME-dev/usr/" >> debian/rules
 #Create some misc files
 mkdir -p debian/source
 echo "8" > debian/compat
